src/generator/requirements.txt
boto3==1.28.62
requests==2.31.0
reportlab==4.0.4
Flask==2.3.3
gevent==23.7.0
python-slugify>=8.0.0
fpdf2>=2.7.8
ydb